All functions and types

LanguageDynamics.AbstractInteractor

LanguageDynamics.AbstractLanguage

LanguageDynamics.AbstractLatticePopulation

LanguageDynamics.AbstractNetworkPopulation

LanguageDynamics.AbstractPopulation

LanguageDynamics.AbstractProbabilisticSpeaker

LanguageDynamics.AbstractSpatialPopulation

LanguageDynamics.AbstractSpeaker

LanguageDynamics.AbstractUtteranceSelector

LanguageDynamics.EmptySpeaker

LanguageDynamics.act!

LanguageDynamics.AbstractInteractor — Type

Abstract interactor. The top node of the Interactor type hierarchy.

LanguageDynamics.AbstractLanguage — Type

Abstract language: an interactor that is a language (community).

LanguageDynamics.AbstractLatticePopulation — Type

Abstract lattice population: interactors on some lattice.

LanguageDynamics.AbstractNetworkPopulation — Type

Abstract network population. Interactors are nodes of some network.

LanguageDynamics.AbstractPopulation — Type

Abstract population. The top node of the Population type hierarchy.

LanguageDynamics.AbstractProbabilisticSpeaker — Type

Abstract probabilistic speaker: a speaker containing a probabilistic replicator.

LanguageDynamics.AbstractSpatialPopulation — Type

Abstract spatial population. Represents any type of population in which interactors have a spatial representation (e.g. a point on a plane).

LanguageDynamics.AbstractSpeaker — Type

Abstract speaker: an interactor that is an individual.

LanguageDynamics.AbstractUtteranceSelector — Type

Abstract utterance selector: an interactor of the Utterance Selection Model type.

LanguageDynamics.EmptySpeaker — Type

An empty speaker. This is an empty type that has no functionality apart from 'existing'. Its use case is testing and debugging population-dynamic functions; linguistically, empty speakers are inert. They can be rendezvoused, technically, but nothing happens.

LanguageDynamics.act! — Method

act!(x::EmptySpeaker, y::AbstractSpeaker)

Make an EmptySpeaker attempt an action on another speaker. Does nothing and returns false.