Auxiliaries

LanguageDynamics.Point — Type

A 2D point of type T; has fields x and y, both of type T. A frequent use case is Point{Int} for a point in the two-dimensional Cartesian product of integers.

LanguageDynamics.chebyshev — Method

chebyshev(x::Point{Int}, y::Point{Int})

Find the Chebyshev distance between two integer points.

LanguageDynamics.manhattan — Method

manhattan(x::Point{Int}, y::Point{Int})

Find the Manhattan distance between two integer points.

LanguageDynamics.zipf — Method

zipf(N::Int, s::Float64)

Zipf's Law for N elements and scaling exponent s, i.e. the probability mass function

\[f(k; s, N) = H_{N,s}^{-1} k^{-s}\]

for $k = 1, \dots , N$, where $H_{N,s}$ is the generalized harmonic number used for normalization.