Interactors

LanguageDynamics.characterize — Method

characterize(x::AbstractInteractor)

Characterize the state of a single interactor. Calls the characterize method on the replicator carried by the interactor.

LanguageDynamics.EmptyInteractor — Type

An empty interactor. This is an empty type that has no functionality apart from 'existing'. Its use case is testing and debugging population-dynamic functions; linguistically, empty interactors are inert. They can be rendezvoused, technically, but nothing happens.

LanguageDynamics.EmptyInteractor — Method

Construct an empty interactor.

LanguageDynamics.act! — Method

act!(x::AbstractInteractor, y::EmptyInteractor)

Make an interactor attempt an action on an EmptyInteractor. Does nothing and returns false.

LanguageDynamics.act! — Method

act!(x::EmptyInteractor, y::AbstractInteractor)

Make an EmptyInteractor attempt an action on another interactor. Does nothing and returns false.

LanguageDynamics.act! — Method

act!(x::EmptyInteractor, y::EmptyInteractor)

Make an EmptyInteractor attempt an action on another EmptyInteractor. Does nothing and returns false.

LanguageDynamics.MomentumSelector — Type

Momentum-based utterance selector, after Stadler et al. (2016).

Reference

Stadler, K., Blythe, R. A., Smith, K. & Kirby, S. (2016) Momentum in language change: a model of self-actuating s-shaped curves. Language Dynamics and Change, 6, 171–198. https://doi.org/10.1163/22105832-00602005

LanguageDynamics.MomentumSelector — Method

MomentumSelector(alpha::Float64, gamma::Float64, b::Float64, T::Int, x0::Float64)

Construct a MomentumSelector with learning rate alpha, short-term memory smoothing factor gamma, momentum bias b, number of utterances per interaction T, and initial frequency x0.

LanguageDynamics.act! — Method

act!(x::MomentumSelector, y::MomentumSelector)

Let one MomentumSelector act upon another, i.e. x speaks and y listens.