Populations

LanguageDynamics.characterize — Method

characterize(x::AbstractPopulation)

Characterize the state of a population. Tries to form either a single number or a vector that reflects the population's average behaviour, depending on the types of replicators carried by the interactions in the population. In practice, this is achieved by calling on characterize functions on the interactors.

LanguageDynamics.characterize_by_location — Method

characterize_by_location(x::AbstractLatticePopulation)

Characterize a lattice population cell by cell.

LanguageDynamics.eject! — Method

eject!(x::AbstractPopulation)

Remove a randomly chosen interactor from a population.

LanguageDynamics.eject! — Method

eject!(x::Int, y::AbstractPopulation)

Remove interactor with ID x from population y.

LanguageDynamics.inject! — Method

inject!(x::AbstractInteractor, y::AbstractPopulation)

Add an interactor to a population.

LanguageDynamics.rendezvous! — Method

rendezvous!(x::AbstractPopulation, y::Int, z::Int)

Conduct a rendezvous between interactors indexed by y and z in population x. (The first mentioned acts first.)

LanguageDynamics.rendezvous! — Method

rendezvous!(x::AbstractPopulation)

Conduct a rendezvous between two interactors chosen at random in population x.

LanguageDynamics.PoolPopulation — Type

A pool population is a well-mixing finite community, i.e. one in which every interactor is connected to every other interactor and the probability of interaction is uniform.

LanguageDynamics.PoolPopulation — Method

PoolPopulation{T}()

Construct an empty pool population of interactors of type T.

Examples

PoolPopulation{EmptyInteractor}()

LanguageDynamics.ZipfTravellerPopulation — Type

A 'Zipfian traveller population': interactors are located on a two-dimensional lattice. Each interactor has a 'home' cell on the lattice, but also has the option of travelling to another cell. The probability to travel to a cell with coordinates $x'$ and $y'$ scales as the inverse of the distance between that cell and the interactor's home (coordinates $x$ and $y$), measured as the Manhattan distance $d$:

\[P \propto \frac{1}{(1 + d)^\lambda}\]

where $\lambda$ is a scaling parameter. Interactors only rendezvous with other interactors currently residing in the same cell.

LanguageDynamics.ZipfTravellerPopulation — Method

ZipfTravellerPopulation(K::Int, lambda::Float64)

Construct a Zipf traveller population of size K by K with scaling parameter lambda.

LanguageDynamics.eject! — Method

eject!(x::Int, y::ZipfTravellerPopulation)

Remove an interactor from a Zipf traveller population.

LanguageDynamics.inject! — Method

inject!(x::AbstractInteractor, y::ZipfTravellerPopulation)

Insert an interactor into a Zipf traveller population.

LanguageDynamics.rendezvous! — Method

rendezvous!(x::ZipfTravellerPopulation)

Rendezvous in a Zipf traveller population (between two random interactors currently in the same cell).

LanguageDynamics.travel! — Method

travel!(x::Int, y::ZipfTravellerPopulation)

Make interactor with index x travel in a Zipf traveller population.

LanguageDynamics.travel! — Method

travel!(x::ZipfTravellerPopulation)

Make a random interactor in a Zipf traveller population travel.

LanguageDynamics.travel_all! — Method

travel_all!(x::ZipfTravellerPopulation)

Make every interactor in a Zipf traveller population travel.